parser/tests/tree_construction/basic.dat

#data
<a><p>X<a>Y</a>Z</p></a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,10): unexpected-start-tag-implies-end-tag
(1,10): adoption-agency-1.3
(1,24): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <a>
|     <p>
|       <a>
|         "X"
|       <a>
|         "Y"
|       "Z"
